python-requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.39.0
celery==3.1.23
Django==1.9.2
django-filter==0.12.0
django-rest-swagger==0.3.5
django-storages==1.4
djangorestframework==3.3.2
drfdocs==0.0.9
facebook-sdk==1.0.0
kombu==3.0.35
Markdown==2.6.5
MySQL-python==1.2.5
python-dateutil==2.5.1
pytz==2016.4
PyYAML==3.11
requests==2.10.0
six==1.10.0
wheel==0.29.0
